pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "django_template"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2.16"
django-extensions = "^3.2.1"
gunicorn = "^20.1.0"
dj-database-url = "^1.2.0"
whitenoise = "^6.2.0"
psycopg2 = "^2.9.5"
sentry-sdk = "^1.12.1"
factory-boy = "^3.2.1"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
flake8 = "^6.0.0"
pre-commit = "^2.21.0"
bandit = "^1.7.4"
detect-secrets = "^1.4.0"
django-types = "^0.16.0"
dslr = "^0.4.0"
djhtml = "^1.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
